Fixed portal, though I wouldn't rely on it. Added some stuff
Commented out the dimension travel thing on enderninja.
---
The custom GUI now works for a fact, look at top of screen for proofs :P
Also cleaned some variable names in it for fun
---
Teleporter now teleports you to the shadow realm, however, don't do it
in creative mode because it makes a nether portal and will instantly
teleport you to the nether. Nether portals seem to instantly teleport
you in creative mode instead of having a delay. :/